Added carousel, modal class, navbar, collapsing extension to navbar
Added a carousel with a bit of a hacky way of making the images behave. Will have to come back and fix that somehow.

Added a bootstrap navbar, and a button that extends a collapsible extension below the navbar with links to my other projects.

Added a modal class as a warning when the user attempts to view my first website.